AG-9402 - Extract chart library url functions
So that they can be used on the frontend without any nodejs imports eg, glob
