+++ packages/ag-charts-website/src/features/example-runner/framework-templates/TypescriptTemplate.astro
@@ -1,10 +1,10 @@
 ---
-import { getCacheBustingUrl } from '@utils/pages';
 import { MetaData } from './lib/MetaData';
 import { ExampleStyle } from './lib/ExampleStyle';
 import { Styles } from './lib/Styles';
 import { SystemJs } from './lib/SystemJs';
 import { pathJoin } from '@utils/pathJoin';
+import { getCacheBustingUrl } from '@utils/chartLibraryPaths';
 
 interface Props {
     isDev: boolean;
