+++ packages/ag-charts-website/src/pages/index.astro
@@ -9,7 +9,7 @@ import HomepageExampleRunner from '@features/homepage/components/HomepageExample
 import { HomepageSeriesSelector } from '@features/homepage/components/HomepageSeriesSelector';
 import { pathJoin } from '@utils/pathJoin';
 import { SITE_BASE_URL } from '@constants';
-import { getCacheBustingUrl, getChartScriptPath, getChartEnterpriseScriptPath } from '@utils/pages';
+import { getCacheBustingUrl, getChartEnterpriseScriptPath, getChartScriptPath } from '@utils/chartLibraryPaths';
 
 const frameworksData = [
     {
