+++ packages/ag-charts-website/src/utils/chartLibraryPaths.ts
@@ -1,0 +1,13 @@
+import { pathJoin } from './pathJoin';
+
+export const getChartScriptPath = (sitePrefix?: string) => {
+    const sitePrefixUrl = sitePrefix ? sitePrefix : '';
+    return pathJoin(sitePrefixUrl, '/dev/ag-charts-community/dist/main.umd.js');
+};
+
+export const getChartEnterpriseScriptPath = (sitePrefix?: string) => {
+    const sitePrefixUrl = sitePrefix ? sitePrefix : '';
+    return pathJoin(sitePrefixUrl, '/dev/ag-charts-enterprise/dist/main.umd.js');
+};
+
+export const getCacheBustingUrl = (url: string, timestamp: number) => `${url}?t=${timestamp}`;
